Move LeafPartitionQueryExecutor to query-core module
