+++ java/query/query-runner/src/main/java/sleeper/query/runner/recordretrieval/QueryExecutor.java
@@ -28,6 +28,7 @@ import sleeper.core.record.Record;
 import sleeper.core.statestore.StateStore;
 import sleeper.core.statestore.StateStoreException;
 import sleeper.core.util.ObjectFactory;
+import sleeper.query.core.LeafPartitionQueryExecutor;
 import sleeper.query.core.LeafPartitionRecordRetriever;
 import sleeper.query.core.model.LeafPartitionQuery;
 import sleeper.query.core.model.Query;
